Move save/load components to their own folder
